chore: Add ActivityResponse model to user module
This commit adds the `ActivityResponse` model to the user module in the backend. The `ActivityResponse` model represents an activity with a date and a count. This model is used to store and retrieve activity data for a user.

The code changes include the addition of the `ActivityResponse` class in the `user.py` file. This class extends the `BaseModel` and defines the `date` and `count` attributes.

Added sortby and order query params